Listener added and @webservlet too so less code in web.xml

+++ maven/src/main/java/hr/excilys/servlet/DashboardServlet.java
@@ -8,13 +8,16 @@ import hr.excilys.mapper.ArrayMapper;
 import hr.excilys.model.Pagination;
 import hr.excilys.service.DashboardService;
 
+import javax.servlet.ServletContext;
 import javax.servlet.ServletException;
+import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.sql.SQLException;
 
+@WebServlet("/dashboard")
 public class DashboardServlet extends HttpServlet {
 
 	private static final long serialVersionUID = 1L;
@@ -23,13 +26,11 @@ public class DashboardServlet extends HttpServlet {
 	private static final String DELETESUCCESS = "1";
 	final Logger logger = LoggerFactory.getLogger(DashboardServlet.class);
 	private DashboardService dashboardService;
-	private Pagination page;
 
 	public DashboardServlet() {
 
 		super();
 		this.dashboardService = new DashboardService();
-		this.page = new Pagination();
 	}
 
 	protected void doGet(HttpServletRequest request, HttpServletResponse response)
@@ -39,7 +40,7 @@ public class DashboardServlet extends HttpServlet {
 				request.getParameter("page"), request.getParameter("search")).order(request.getParameter("order"))
 						.direction(request.getParameter("direction")).build();
 
-		page = dashboardService.paginate(dtoPagination);
+		Pagination page = dashboardService.paginate(dtoPagination);
 
 		request.setAttribute("ok", request.getParameter("ok"));
 		request.setAttribute("msg", request.getParameter("msg"));
@@ -51,23 +52,39 @@ public class DashboardServlet extends HttpServlet {
 		request.setAttribute("size", page.getCount());
 		request.setAttribute("order", page.getOrder());
 		request.setAttribute("direction", page.getDirection());
-		request.setAttribute("computers", dashboardService.getComputersRows(page));
 
-		this.getServletContext().getRequestDispatcher("/WEB-INF/dashboard.jsp").forward(request, response);
+		ServletContext servletContext = getServletContext();
+		Object obj = servletContext.getAttribute("DashboardService");
+		if (obj instanceof DashboardService) {
+			DashboardService dashboardService = (DashboardService) obj;
+			request.setAttribute("computers", dashboardService.getComputersRows(page));
+			this.getServletContext().getRequestDispatcher("/WEB-INF/dashboard.jsp").forward(request, response);
+		} else {
+			throw new ServletException("dashboardService unavailable");
+		}
 	}
 
 	protected void doPost(HttpServletRequest request, HttpServletResponse response)
 			throws ServletException, IOException {
 
-		try {
-			if (!dashboardService.deleteComputer(ArrayMapper.stringTransform(request.getParameter("selection")))) {
-				request.setAttribute("msg", ARRAYEMPTY);
-				this.getServletContext().getRequestDispatcher("/WEB-INF/dashboard.jsp").forward(request, response);
+		ServletContext servletContext = getServletContext();
+		Object obj = servletContext.getAttribute("DashboardService");
+		if (obj instanceof DashboardService) {
+			DashboardService dashboardService = (DashboardService) obj;
+			try {
+				if (!dashboardService.deleteComputer(ArrayMapper.stringTransform(request.getParameter("selection")))) {
+					request.setAttribute("msg", ARRAYEMPTY);
+					doGet(request, response);
+				}
+			} catch (NumberFormatException | SQLException e) {
+				request.setAttribute("msg", DELETEERROR);
+				doGet(request, response);
 			}
-		} catch (NumberFormatException | SQLException e) {
-			request.setAttribute("msg", DELETEERROR);
-			this.getServletContext().getRequestDispatcher("/WEB-INF/dashboard.jsp").forward(request, response);
+			request.setAttribute("msg", DELETESUCCESS);
+			doGet(request, response);
+		} else {
+			throw new ServletException("dashboardService unavailable");
 		}
-		response.sendRedirect("?msg=" + DELETESUCCESS);
+
 	}
 }
